[itests] update on raft_consensus_stress-itest
Updates on raft_consensus_stress-itest after running it in 3-2-3 mode.

I found that sometimes 30 iterations of ksck was not enough to converge
if running at dist-test machines with the 3-2-3 re-replication mode,
but 50 iterations plus delay of one heartbeat interval between retries
was enough to make it more stable for the 3-2-3 mode runs.

Change-Id: I6762f1783f4e926225befe3e49f4e12b58a27dd1
Reviewed-on: http://gerrit.cloudera.org:8080/9567
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>