[consensus] FAILED_UNRECOVERABLE replica health status
Added HealthStatus::FAILED_UNRECOVERABLE for a tablet replica. This is
to mark replicas which are not able to catch up with the leader due to
GC-collected segments of WAL and other unrecoverable cases.

With the introduction of the FAILED_UNRECOVERABLE health status, the
replica management scheme becomes hybrid: the system evicts replicas
with FAILED_UNRECOVERABLE health status before adding a replacement
if it anticipates that it can commit the transaction.

This patch is a part of the fix to address KUDU-2342.  It also addresses
KUDU-2322 as well: evicting voter replicas more aggressively if they
fall behind log segment GC threshold.

Change-Id: I35637c5bda6681b732dbc2bbf94b9d4258b12095
Reviewed-on: http://gerrit.cloudera.org:8080/9625
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>
(cherry picked from commit a74f9a0dcaf88315c8563b95cdeb5701d9ce5438)
Reviewed-on: http://gerrit.cloudera.org:8080/9649
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@gmail.com>