KUDU-2346: Fix mac pkg-config issue.
Previously an empty return string from pkg-config --cflags was used
as a proxy for testing if the package was installed. The MacOs version
of pkg-config returns an error message in response to this command
when a package is not installed.

The fix is to property test the return value of pkg-config --cflags.

Change-Id: I8a6b543dc65ac65d0dedf7e39f735b4805612bf1
Reviewed-on: http://gerrit.cloudera.org:8080/9652
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>