KUDU-16 pt 1: add server-side limits on scanners
This patch adds the functionality to specify limits to the number of
rows returned per tablet server scanner. Internally, each server-side
scanner now tracks the number of rows it has already returned and
adjusts the number of rows to serialize per batch based on this limit.
Setting a non-positive limit is grounds for short circuiting the scan.

Note: Use of this functionality by the client will come in a later
patch.

Change-Id: I3ca5ddff09f4910d12f80259bd11e4027f99aa7c
Reviewed-on: http://gerrit.cloudera.org:8080/9783
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>