[rpc-test] macOS-specific fix for TestRpcSidecarLimits
As Will noticed, after TestRpcSidecarLimits was updated with 2b0c1c0,
on macOS the following new error surfaced while running the test
(which is completely legit in accordance with 'man sendmsg'):

  Network error: sendmsg error: Invalid argument (error 22)

This patch updates the pattern of the expected error message for the
case if sendmsg() returns EINVAL when the sum of the iov_len values
overflows an ssize_t.

This is a follow-up for 2b0c1c019921e485f06c4be280fedba3d5279672.

Change-Id: I17681546fd942724025dfde81eaa412ee7e0acbf
Reviewed-on: http://gerrit.cloudera.org:8080/9719
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>