[delete_table-itest] fix flake in TestNoDeleteTombstonedTablets
Fixed flake in DeleteTableITest.TestNoDeleteTombstonedTablets scenario.
The flake started to surface after the 3-4-3 replica management scheme
was turned on by default.

In essence, once AddServer() was called, it might happen that the
following RemoveServer() would race with the catalog manager trying
to remove an excess replica, resulting in an error message like below:

  src/kudu/integration-tests/delete_table-itest.cc:1212: Failure
  Failed
  Bad status: Illegal state: RaftConfig change currently pending. \
      Only one is allowed at a time.

I verified the fix running the scenario using dist_test:
  prior to the fix (12 out of 256 failed)
    http://dist-test.cloudera.org/job?job_id=aserbin.1521584964.136771

  after the fix (0 out of 256 failed)
    http://dist-test.cloudera.org//job?job_id=aserbin.1521586527.79020

Change-Id: I9fddc0667b22d258385b9cd7ec2ed52fd5e052d8
Reviewed-on: http://gerrit.cloudera.org:8080/9729
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>