tools: return error when updating with no block manager instances
Previously when running `kudu fs update_dirs` with a set of directories
that did not contain existing block manager instances, the tool would
fail a CHECK.

Prior to support for configurable metadata directories (6492633), the
CHECK was guaranteed, as at least the first data directory was
guaranteed to exist (or an error would be returned earlier on
otherwise). Now, rather than failing a CHECK, we should return an error.
If Kudu can't find any block manager instances, the following error is
returned instead:

Not found: cannot update data directories: could not open directory manager; none of the provided data directories could be found.

Note that this only affects the `kudu fs update_dirs` tool; "normal"
opening of the FS layout would return errors if any block manager
instances are not found.

Change-Id: I06d95f561301eebc21542e01782a8c219ca71c70
Reviewed-on: http://gerrit.cloudera.org:8080/9737
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>