pprof: fix /pprof/contention to work with the go version of pprof
The old perl version of pprof that ships with gperftools is no longer
maintained. It's been replaced by github.com/google/pprof which is
written in go.

The go version is slightly pickier about the output than the old perl
version. This patch updates a few things so that the new version works
with the /pprof/contention profile:

- the expected first line format is slightly different
- hex addresses must be prefixed with '0x', no tab characters expected
- memory maps needs a section header ahead of it
- cycles/second needs to be formatted as an integer, not float

I verified that with this patch I was able to run the following:

  pprof -seconds=1  http://localhost:8050/pprof/contention

... and got a reasonable looking profile when running under load.

It also appears that the old perl pprof is still forgiving enough to
handle this new format.

Change-Id: Ie25b2332688575c57be10b437566adb3f0c156e3
Reviewed-on: http://gerrit.cloudera.org:8080/9741
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>