KUDU-2356. Idle WALs should not consume significant memory
Prior to this patch, each WAL writer would keep its temporary
compress_buf_ sized based on the largest data that was ever written to
it, even if the tablet was idle. Keeping the buffer around for a hot WAL
probably has some small performance benefit, but when the WAL is idle
enough to shut down its thread it should also clear its buffer to avoid
occupying memory.

Change-Id: I977976ef9505401b1bd596046915473a91153cd9
Reviewed-on: http://gerrit.cloudera.org:8080/9747
Tested-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>