KUDU-2075. Fix SEGV when using tracing from vanilla threads
This fixes a crash when tracing was enabled and a thread happened to
reuse the same pthread_self identifier as a previous deceased
non-kudu::Thread-based thread.

The issue was that non-Kudu-threads were not unregistering themselves
from the TraceLog instance, so when the new thread started and generated
a trace event, it would fail with a CHECK failure like:

F0721 13:14:12.038748  2708 map-util.h:414] Check failed:
InsertIfNotPresent(collection, key, data) duplicate key: 139914842822400

The new test reproduced the issue reliably.

This patch also removes Thread::CallAtExit() which was no longer used
after removing this callsite.

Change-Id: I7457be81c1c2668244907a70d62db7cf91321375
Reviewed-on: http://gerrit.cloudera.org:8080/9750
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins