KUDU-972. Cache should track the memory overhead
This changes the cache 'charge' mechanism to use malloc_usable_size on
its actual allocations rather than tracking only the cached _values_.
This addresses two sources of inaccuracy:

1) we weren't counting the cache key and LRUHandle overhead, which is
apparently 104 bytes per object.

2) we weren't accounting for tcmalloc internal fragmentation, which can
be significant for allocations in our typical block size range. For
example, a 4097-byte allocation will use 4608 bytes (12% overhead). In
an insert-heavy workload most of hte cache is used for bloom blocks
which are ~4kb, so the above is a very common case.

Change-Id: I6cc69004dccb1cc372fdedbac321154e46a92263
Reviewed-on: http://gerrit.cloudera.org:8080/9752
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins