[consensus] KUDU-2367 fix replica health report
This patch addresses KUDU-2367,where a leader replica under particular
conditions reports follower's health status as FAILED instead of
FAILED_UNRECOVERABLE. I.e., the bug in the code of the
PeerMessageQueue::PeerHealthStatus() method resulted in having the
leader reporting the health status of the replica still as FAILED even
after the replica fell behind the WAL segment GC threshold.

Also, a couple of already existing integration tests are updated to
provide appropriate coverage for this patch.

Change-Id: If77474adc8f618f2cda35f992190133138fdb511
Reviewed-on: http://gerrit.cloudera.org:8080/9755
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins