consensus: Expose health status in GetConsensusState() RPC
This patch exposes health reporting as part of the GetConsensusState()
RPC method by plumbing in an additional member to the RPC request. This
is useful for testing purposes.

The helper function in cluster_itest_util.h is also updated, as well as
all outstanding call sites.

A test that uses this new functionality is in a follow-up patch.

Change-Id: I15b8931156183e1713b902d22abbc80ca393f6b9
Reviewed-on: http://gerrit.cloudera.org:8080/9781
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>