KUDU-16 pt 2: add client-side limits on scanners
This patch adds a public API to allow the specification of
per-client-side-scanner limits on the number of rows returned. Each
scanner will maintain a count of the number of rows already read, and
adjust the server-side limit upon sending the next scan request.

A couple of tests are included to verify that the limits act as
expected. I also verified that lowering the limit reduces the number of
bytes read on disk (at the granularity of a single scan batch at a
time).

Note: this patch does not implement the behavior for scan tokens.

Change-Id: Ib2d40e3d14e36f3bf1d09a4bfdb3e17a745d244d
Reviewed-on: http://gerrit.cloudera.org:8080/9790
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>