consensus: Revert terminal state enforcement of FAILED_UNRECOVERABLE
In patch 7b88a42332540ace18ea033439ceac0e7cb68411 FAILED_UNRECOVERABLE
was noted as a terminal state. Todd pointed out that there are certain
corner cases where that may not be the case, such as a replica that
started up on a temporarily-failed or read-only disk, and then was
restarted. In such a case, if the leader was previously unable to evict
that node because of a lack of a majority, the affected replica may
revert to a HEALTHY state.

Change-Id: I2f858af8c578340f7a32b95702b5da7bbe81a675
Reviewed-on: http://gerrit.cloudera.org:8080/9794
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins