[docs] Fix bad subsection and relocate "Disk Issues" section
The section on rebuilding a Kudu filesystem layout was improperly
a subsection of the majority recovery section.

The performance section of the troubleshooting page had a "Disk Issues"
section describing a failure to startup due to a partial Kudu fs or a
non-empty initial set of fs directories. I moved this to be a child
section of the "Startup Errors" section. I changed the title as well,
but left the id the same so old links will still work.

Change-Id: Ia2e2abc1359e0a0fd968fe1a85073714af1ee695
Reviewed-on: http://gerrit.cloudera.org:8080/9816
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>