Introduce examples directory
This introduces the top-level examples directory, which will be for
examples of how to use the Kudu clients and integrations.

As a first step, I moved the client examples from src/kudu/client/examples
to examples/cpp. I also renamed "sample" -> "example" generally.

I also manually verified that the instructions in the README for building the
example application still work.

Change-Id: Iceca48771541f95e4f0caf8c098bc85d34aa558e
Reviewed-on: http://gerrit.cloudera.org:8080/9837
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>