[java] Fix javadoc issues
Fixes some javadoc warnings:
- Removed @link and @see to external classes
- Remove invalid @inheritDoc tags
- Fixed code block
- Fixed a couple @return statment

This is far from all of the correctness issue, but the few
that were resulting in Javadoc Gradle errors.

Also disabled the super strict doc linter for Java 8+
in the gradle build to prevent a ton of warning output
and potnetial build failures. We could enable this at
somepoint after we correct the hundreds of warnings.

Change-Id: I5914b416e1b084a4f50dd8ea878be2f065acbe94
Reviewed-on: http://gerrit.cloudera.org:8080/9839
Tested-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>