KUDU-2394. Fix flaky RpcStubTest.TestDontHandleTimedOutCalls
This test was flaky under stress+ASAN with three cases of failures:

(1) most commonly, the assertion to check that an RPC had timed out in the
queue would run too quickly, just before the actual metric was incremented.
Changing it to ASSERT_EVENTUALLY got rid of most of the failures.

(2) in a few cases, the 100ms sleeps were processed completely before the
later RPC was processed, so the RPC that was supposed to time out in the
queue actually had a chance to run. Switching to 1000ms sleeps got rid
of the remaining failure case.

(3) in a few other cases, the call that was supposed to time out (with a
1ms timeout) timed out before it was even sent to the server. This patch bumps
the timeout to 5ms and retries that until it succeeds with a server-side
timeout.

Ran 5000x with the following command:

./build-support/dist_test.py loop -n 500 build/latest/bin/rpc_stub-test \
  --gtest_filter=\*DontHandle\* --gtest_repeat=10 --stress-cpu-threads=8

and got no failures after this patch, whereas I'd usually get 5-10 failures
prior.

Change-Id: If8cfebb5e9160fb11586abdbba6fc1ae06574a2c
Reviewed-on: http://gerrit.cloudera.org:8080/9897
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>