[tools] ksck improvements [4/n]: KUDU-2271 and KUDU-2310
This improves formatting in ksck output and ensures that the tablet ids
appear in the output in verbose mode (KUDU-2271).

It also revamps how results are printed when tablet id filters are
specified, omitting conclusions about entire tables. It's now clear from
ksck's messages whether it is checking individual tablets using
tablet filters or checking whole tables. This addresses KUDU-2310 as
well, since it makes it clear what sort of filters are being applied.
The summary table is unchanged since it's nice whether a table or tablet
filter is applied.

Change-Id: I2371b704a7c2ebb1b5a4822e3b4d852443c200c5
Reviewed-on: http://gerrit.cloudera.org:8080/9912
Reviewed-by: Attila Bukor <abukor@cloudera.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>