perf: avoid unnecessary floating point division and multiplication
I noticed looking at a profile that Trace::SubstituteAndTrace was
spending a fair number of cycles in some floating point math, and
tracked it down to our use of constants like '1e3' and '1e6'. These are
floating point constants even though we would do just as well with
integer multiplication and division for this use case.

I verified that switching to the integer versions of the constants
changed the generated assembly to avoid costly division instructions[1].

This patch fixes that case as well as a few others with the same
pattern.

[1] https://godbolt.org/g/dmE7Av

Change-Id: I3985433056aab163cc65972e94bc2108f00546a3
Reviewed-on: http://gerrit.cloudera.org:8080/9894
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>